% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assert_valid_tidy_pairwise_simdist}
\alias{assert_valid_tidy_pairwise_simdist}
\title{Assert Input is a valid of tidy pairwise similarity/dissimilarity data.frame}
\usage{
assert_valid_tidy_pairwise_simdist(data)
}
\arguments{
\item{data}{A dataframe containing pairwise similarity or dissimilarity data.}
}
\description{
This function checks the validity of a dataframe representing pairwise similarity or
dissimilarity data in a tidy format. It ensures that the input dataframe has at least
three columns representing subject identifiers and their pairwise similarity or distance values.
Additionally, it checks that all required columns have non-missing numeric values.
}
\examples{
# Create a tidy pairwise simdist dataframe
tidy_simdist_data <- tidy_pairwise_jaccard_similarity(my_data)

# Check the validity of the tidy pairwise simdist dataframe
assert_valid_tidy_pairwise_simdist(tidy_simdist_data)

}
\seealso{
\code{\link{tidy_pairwise_jaccard_similarity}}
}
