% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{family}
\alias{family}
\title{Smith vs Johnson Family}
\format{
\subsection{\code{family}}{

A data frame with 55 rows and 2 variables:
\describe{
\item{name}{name of person}
\item{trait}{physical trait of the person}
...
}
}
}
\source{
Simulated
}
\usage{
family
}
\description{
Simulated dataset describing physical traits individual members of two different families
}
\details{
\itemize{
\item Family 1: Smith (5 members)
\item Family 2: Johnson (3 members)
}

Each row represents one physical feature possessed by one person.
Each person can appear numerous times since each person has multiple physical features.
}
\keyword{datasets}
