% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{family_member_annotations}
\alias{family_member_annotations}
\title{Smith vs Johnson Family}
\format{
\subsection{\code{family}}{

A data frame with 8 rows and 4 variables:
\describe{
\item{name}{name of person}
\item{family}{which family does this person belong to}
\item{traits}{comma-delimited list of physical traits}
\item{tooltip}{text that we'll use as a tooltip for leaf nodes}
...
}
}
}
\source{
Simulated
}
\usage{
family_member_annotations
}
\description{
Dataset describing metadata about individual members of two different families.
Used to change colours, shapes, etc.  in dendograms
}
\details{
\itemize{
\item Family 1: Smith (5 members)
\item Family 2: Johnson (3 members)
}

Each person appears in the dataset exactly once
}
\keyword{datasets}
