% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard Index Calculation}
\usage{
jaccard(a, b)
}
\arguments{
\item{a}{A vector representing the first set.}

\item{b}{A vector representing the second set.}
}
\value{
A numeric value representing the Jaccard index between the two sets.
}
\description{
Calculate the Jaccard index between two sets.
}
\details{
The Jaccard index, also known as the Jaccard similarity coefficient, measures the
similarity between two sets by comparing the size of their intersection to the size
of their union.
}
\examples{
jaccard(c(1, 2, 3), c(2, 3, 4))

jaccard(factor(c(1, 2, 3)), factor(c(2, 3, 4)))

jaccard(c(1), c())

}
\seealso{
\code{\link{intersect}} calculates the intersection of two vectors.
\code{\link{length}} calculates the length of a vector.
}
