% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matrix_to_dist_class}
\alias{matrix_to_dist_class}
\title{Convert Matrix to 'dist' Class object}
\usage{
matrix_to_dist_class(matrix)
}
\arguments{
\item{matrix}{A numeric matrix.}
}
\value{
A distance class object.
}
\description{
NOT to be confused with \code{\link[=matrix_toggle_simdist]{matrix_toggle_simdist()}} which converts similarity matrices to distance matrices and vice versa.
}
\details{
This function converts a matrix to a 'dist' class object using the \code{\link[stats]{as.dist}} function.
This 'dist' class object can be fed into hclust and other functions
}
\examples{
matrix <- matrix(c(1, 2, 3, 2, 0, 4, 3, 4, 0), ncol = 3)
result <- matrix_to_dist_class(matrix)

}
\seealso{
\code{\link[stats]{as.dist}}
}
