% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matrix_toggle_simdist}
\alias{matrix_toggle_simdist}
\title{Convert Similarity Matrix to Distance Matrix and Vice Versa}
\usage{
matrix_toggle_simdist(matrix)
}
\arguments{
\item{matrix}{A numeric matrix representing either a similarity or a distance matrix.}
}
\value{
A matrix of the same dimensions as the input matrix, converted to the opposite type.
}
\description{
This function takes a matrix representing either a similarity or a distance matrix
and converts it to its counterpart. For similarity matrices, it subtracts each value
from 1 to obtain a distance matrix. For distance matrices, it performs the reverse
operation to obtain a similarity matrix.
}
\examples{
# Create a similarity matrix
sim_matrix <- matrix(c(1.0, 0.8, 0.5, 0.8, 1.0, 0.6, 0.5, 0.6, 1.0), nrow = 3)

# Convert similarity matrix to distance matrix
dist_matrix <- matrix_toggle_simdist(sim_matrix)

}
