% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tidy_pairwise_sample_combinations}
\alias{tidy_pairwise_sample_combinations}
\title{Generate Tidy Pairwise Sample Combinations}
\usage{
tidy_pairwise_sample_combinations(
  samples,
  prefix = "sample",
  include_unused_levels = TRUE
)
}
\arguments{
\item{samples}{A character vector containing the list of samples.}

\item{prefix}{A character string to be used as a prefix for the column names in the output dataframe.
By default columns are named 'sample1' and 'sample2'.}

\item{include_unused_levels}{if samples is a factor with unused levels, should those levels be included in the pairwise matrix?}
}
\value{
A 2-column dataframe with representing unique pairwise combinations of samples.
}
\description{
This function takes a vector of samples and returns a tidy dataframe with every unique
pairwise combination of samples.
}
\examples{

samples <- c("A", "B", "C")
tidy_pairwise_sample_combinations(samples)


}
\seealso{
\href{https://www.rdocumentation.org/packages/utils/versions/3.6.2/topics/combn}{combn function documentation}
}
