% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{visualisation_make_interactive}
\alias{visualisation_make_interactive}
\title{Make a ggplot2 Visualization Interactive}
\usage{
visualisation_make_interactive(gg, width_svg = 6, height_svg = 3, ...)
}
\arguments{
\item{gg}{A ggplot2 plot object to be made interactive.}

\item{width_svg}{Width of the SVG container for the interactive plot.}

\item{height_svg}{Height of the SVG container for the interactive plot.}

\item{...}{
  Arguments passed on to \code{\link[ggiraph:girafe]{ggiraph::girafe}}
  \describe{
    \item{\code{code}}{Plotting code to execute}
    \item{\code{ggobj}}{ggplot object to print. Argument \code{code} will
be ignored if this argument is supplied.}
    \item{\code{pointsize}}{the default pointsize of plotted text in pixels, default to 12.}
    \item{\code{width_svg,height_svg}}{The width and height of the graphics region in inches.
The default values are 6 and 5 inches. This will define the aspect ratio of the
graphic as it will be used to define viewbox attribute of the SVG result.

If you use \code{girafe()} in an 'R Markdown' document, we
recommend not using these arguments so that the knitr
options \code{fig.width} and \code{fig.height} are used instead.}
    \item{\code{options}}{a list of options for girafe rendering, see
\code{\link[ggiraph:opts_tooltip]{opts_tooltip()}}, \code{\link[ggiraph:opts_hover]{opts_hover()}}, \code{\link[ggiraph:opts_selection]{opts_selection()}}, ...}
    \item{\code{dependencies}}{Additional widget HTML dependencies, see \code{\link[htmlwidgets:createWidget]{htmlwidgets::createWidget()}}.}
  }}
}
\value{
An interactive ggiraph plot object.
}
\description{
This function takes a ggplot2 plot and makes it interactive using the ggiraph package.
}
