% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{visualise_dendrogram_ggraph}
\alias{visualise_dendrogram_ggraph}
\title{Visualize Dendrogram using ggraph}
\usage{
visualise_dendrogram_ggraph(
  hclust,
  annotations = NULL,
  col_label = NULL,
  col_sample = NULL,
  col_colour = NULL,
  col_shape = NULL,
  col_tooltip = NULL,
  col_data_id = NULL,
  col_classification = NULL,
  circular = FALSE,
  label_leaf_only = FALSE,
  draw_labels = FALSE,
  padding_bottom = 0.6,
  legend_position = "bottom",
  leaf_node_size = 1
)
}
\arguments{
\item{hclust}{The hierarchical clustering object (class 'hclust') to visualize.}

\item{annotations}{A dataframe containing node annotations.}

\item{col_label}{The name of the column in the annotations dataframe that contains node labels.}

\item{col_sample}{The name of the column in the annotations dataframe to be used for sampling.}

\item{col_colour}{The name of the column in the annotations dataframe that contains color information.}

\item{col_shape}{The name of the column in the annotations dataframe that contains shape information.}

\item{col_tooltip}{The name of the column in the annotations dataframe that contains tooltip information.}

\item{col_data_id}{The name of the column in the annotations dataframe that contains data IDs.}

\item{col_classification}{The name of the column in the annotations dataframe that contains classification information.}

\item{circular}{A logical value indicating whether to create a circular dendrogram.}

\item{label_leaf_only}{A logical value indicating whether to label only leaf nodes.}

\item{draw_labels}{A logical value indicating whether to}

\item{padding_bottom}{How much padding to add to the bottom of the plot (used to make space for text)}

\item{legend_position}{Position of the legend in the plot.}

\item{leaf_node_size}{Size of the leaf nodes in the plot.}
}
\value{
A ggraph plot object.
}
\description{
This function creates a visual representation of a dendrogram using the ggraph package.
}
